🏗️ refactor(api): clean up unused exports and improve module organization
- Remove unused exports identified by knip static analysis
- Consolidate imports to use barrel exports from index.ts files
- Rename User to AuthUser in auth context for clarity
- Remove unused functions: safeJson, getAuthProviderConfig, validateAuthProviderConfig, getPool, buildPoolConfig, randomChoice, isResult
- Fix circular import issues by organizing exports through domain index files
- Improve module boundaries across auth, question, quiz, and user domains
- Update all import paths to use consolidated exports
- Remove duplicate type exports and consolidate in appropriate modules
