+++ apps/api/src/features/question/domain/entities/Question.test.ts
@@ -3,7 +3,7 @@
  * @fileoverview Tests for the Question entity domain model
  */
 
-import type { QuestionId } from '@api/features/quiz/domain/value-objects/Ids';
+import type { QuestionId } from '@api/features/quiz/domain';
 import { describe, expect, it } from 'vitest';
 import { QuestionOption } from '../value-objects/QuestionOption';
 import { QuestionOptions } from '../value-objects/QuestionOptions';
