+++ apps/api/src/features/question/domain/entities/Question.ts
@@ -3,7 +3,7 @@
  * @fileoverview Domain entity representing a question in the certification exam system
  */
 
-import { QuestionId } from '@api/features/quiz/domain/value-objects/Ids';
+import { QuestionId } from '@api/features/quiz/domain';
 import { ValidationError } from '@api/shared/errors';
 import { Result } from '@api/shared/result';
 import type { QuestionSummary } from '../repositories/IQuestionRepository';
