+++ apps/api/src/features/question/list-questions/handler.test.ts
@@ -3,7 +3,7 @@
  * @fileoverview Tests for list questions business logic including data quality logging
  */
 
-import { QuestionId } from '@api/features/quiz/domain/value-objects/Ids';
+import { QuestionId } from '@api/features/quiz/domain';
 import type { LoggerPort } from '@api/shared/logger';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import type {
