+++ apps/api/src/features/quiz/domain/index.ts
@@ -14,19 +14,12 @@ export { AggregateRoot } from './base/AggregateRoot';
 // Entities
 export { Answer } from './entities/Answer';
 export { DomainEvent } from './events/DomainEvent';
-// Events
-export {
-  AnswerSubmittedEvent,
-  QuizCompletedEvent,
-  QuizExpiredEvent,
-  QuizStartedEvent,
-} from './events/QuizEvents';
 export { DrizzleQuizRepository } from './repositories/DrizzleQuizRepository';
 // Repositories
 export type { IQuizRepository } from './repositories/IQuizRepository';
 export type { Category, Difficulty, ExamType } from './value-objects/ExamTypes';
 // Value Objects
-export { type OptionId, type QuestionId, QuizSessionId, type UserId } from './value-objects/Ids';
+export { type OptionId, QuestionId, QuizSessionId, type UserId } from './value-objects/Ids';
 export { QuestionOrder } from './value-objects/QuestionOrder';
 export { QuestionReference } from './value-objects/QuestionReference';
 export { QuizConfig } from './value-objects/QuizConfig';
