+++ apps/api/src/infra/auth/AuthProviderFactory.ts
@@ -62,42 +62,3 @@ export function createAuthProvider(config?: Partial<AuthProviderConfig>): IAuthP
       throw new AppError('Unknown auth provider specified', 'UNKNOWN_AUTH_PROVIDER', 500);
   }
 }
-
-/**
- * Get current auth provider configuration from environment
- */
-export function getAuthProviderConfig(): AuthProviderConfig {
-  const provider = getAuthProviderType();
-
-  const config: AuthProviderConfig = {
-    provider,
-  };
-
-  if (provider === 'keycloak') {
-    config.keycloak = {
-      url: process.env.KEYCLOAK_URL || 'http://localhost:8080',
-      realm: process.env.KEYCLOAK_REALM || 'certquiz',
-      clientId: process.env.KEYCLOAK_CLIENT_ID || 'certquiz-api',
-      clientSecret: process.env.KEYCLOAK_CLIENT_SECRET || undefined,
-    };
-  }
-
-  return config;
-}
-
-/**
- * Validate auth provider configuration
- */
-export function validateAuthProviderConfig(config: AuthProviderConfig): void {
-  if (config.provider === 'keycloak') {
-    if (!config.keycloak?.url) {
-      throw new AppError('KeyCloak URL is required', 'KEYCLOAK_URL_REQUIRED', 500);
-    }
-    if (!config.keycloak?.realm) {
-      throw new AppError('KeyCloak realm is required', 'KEYCLOAK_REALM_REQUIRED', 500);
-    }
-    if (!config.keycloak?.clientId) {
-      throw new AppError('KeyCloak client ID is required', 'KEYCLOAK_CLIENT_ID_REQUIRED', 500);
-    }
-  }
-}
