+++ apps/api/src/infra/auth/JwtVerifier.ts
@@ -1,4 +1,4 @@
-import { UserRole } from '@api/features/auth/domain/value-objects/UserRole';
+import { UserRole } from '@api/features/auth/domain';
 import type { AuthUser } from '@api/middleware/auth/auth-user';
 import { createRemoteJWKSet, jwtVerify } from 'jose';
 import type { IRoleMapper } from './RoleMapper';
