+++ apps/api/src/infra/auth/RoleMapper.test.ts
@@ -1,4 +1,4 @@
-import { UserRole } from '@api/features/auth/domain/value-objects/UserRole';
+import { UserRole } from '@api/features/auth/domain';
 import { describe, expect, it } from 'vitest';
 import { RoleMapper } from './RoleMapper';
