+++ apps/api/src/infra/db/client.ts
@@ -67,7 +67,7 @@ function initializeDatabase(): { pool: postgres.Sql; db: DB } {
  * Get the database connection pool
  * Initializes on first use
  */
-export function getPool(): postgres.Sql {
+function getPool(): postgres.Sql {
   const { pool } = initializeDatabase();
   return pool;
 }
