+++ apps/api/src/infra/repositories/providers.ts
@@ -1,7 +1,7 @@
-import type { IUserRepository as IAuthUserRepository } from '@api/features/auth/domain/repositories/IUserRepository';
-import type { IQuestionRepository } from '@api/features/question/domain/repositories/IQuestionRepository';
-import type { IQuizRepository } from '@api/features/quiz/domain/repositories/IQuizRepository';
-import type { IUserRepository } from '@api/features/user/domain/repositories/IUserRepository';
+import type { IUserRepository as IAuthUserRepository } from '@api/features/auth/domain';
+import type { IQuestionRepository } from '@api/features/question/domain';
+import type { IQuizRepository } from '@api/features/quiz/domain';
+import type { IUserRepository } from '@api/features/user/domain';
 import type { TransactionVariables } from '@api/middleware/transaction';
 import type { Context } from 'hono';
