+++ apps/api/src/shared/random/index.ts
@@ -5,6 +5,5 @@
 
 export {
   generateCryptoSeed,
-  randomChoice,
   shuffleWithSeed,
 } from './seeded-rng';
